+++ azext_iot/services/auth.py
